+++ NadekoBot.Core/Services/Database/Repositories/Impl/DiscordUserRepository.cs
@@ -72,9 +72,9 @@ VALUES ({userId}, {username}, {discrim}, {avatarId});
         public long GetUserCurrency(IUser user) =>
             GetOrCreate(user).CurrencyAmount;
 
-        public void RemoveFromMany(List<long> ids)
+        public void RemoveFromMany(List<ulong> ids)
         {
-            var items = _set.Where(x => ids.Contains((long)x.UserId));
+            var items = _set.Where(x => ids.Contains(x.UserId));
             foreach (var item in items)
             {
                 item.CurrencyAmount = 0;
